:truck: extract the app loader and rename register.ts
