+++ src/index.ts
@@ -3,7 +3,7 @@
  * @since 2019-04-25
  */
 
-export { LifeCycles, registerMicroApps, start } from './register';
+export { registerMicroApps, start } from './apis';
 export * from './errorHandler';
 export * from './effects';
 export * from './interfaces';
